+++ test/rules/test_plugin_manager.py
@@ -1313,6 +1313,8 @@ def test_markdown_with_plugins_list_only(scanner_default: MarkdownScanner) -> No
   md049   emphasis-style                  True       True       0.5.0    No
   md050   strong-style                    True       True       0.5.0    No
   md051   link-fragments                  True       True       0.5.0    No
+  md053   link-image-reference-definitio  True       True       0.5.0    No
+          ns
   md059   descriptive-link-text           True       True       0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1399,6 +1401,8 @@ def test_markdown_with_plugins_list_only_all(scanner_default: MarkdownScanner) -
   md049   emphasis-style                  True       True       0.5.0    No
   md050   strong-style                    True       True       0.5.0    No
   md051   link-fragments                  True       True       0.5.0    No
+  md053   link-image-reference-definitio  True       True       0.5.0    No
+          ns
   md059   descriptive-link-text           True       True       0.5.0    No
   md999   debug-only                      False      False      0.0.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
@@ -1487,6 +1491,8 @@ def test_markdown_with_plugins_list_after_command_line_disable_all_rules(
   md049   emphasis-style                  True       False      0.5.0    No
   md050   strong-style                    True       False      0.5.0    No
   md051   link-fragments                  True       False      0.5.0    No
+  md053   link-image-reference-definitio  True       False      0.5.0    No
+          ns
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1579,6 +1585,8 @@ def test_markdown_with_plugins_list_after_configuration_disable_all_rules(
   md049   emphasis-style                  True       False      0.5.0    No
   md050   strong-style                    True       False      0.5.0    No
   md051   link-fragments                  True       False      0.5.0    No
+  md053   link-image-reference-definitio  True       False      0.5.0    No
+          ns
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No
@@ -1673,6 +1681,8 @@ def test_markdown_with_plugins_list_after_command_line_disable_all_rules_and_ena
   md049   emphasis-style                  True       False      0.5.0    No
   md050   strong-style                    True       False      0.5.0    No
   md051   link-fragments                  True       False      0.5.0    No
+  md053   link-image-reference-definitio  True       False      0.5.0    No
+          ns
   md059   descriptive-link-text           True       False      0.5.0    No
   pml100  disallowed-html                 False      False      0.6.0    No
   pml101  list-anchored-indent            False      False      0.6.0    No